+++ src/common/config_parser.cpp
@@ -572,12 +572,29 @@ void ConfigParser::addOptionsTraining(cli::CLIWrapper& cli) {
       "Dynamic cost scaling for mixed precision training: "
       "scaling factor, frequency, multiplier, minimum factor")
       ->implicit_val("8.f 10000 1.f 8.f");
+  
   cli.add<std::vector<std::string>>("--throw-on-divergence",
       "Throw exception if training diverges. Divergence is detected if the running average loss over arg1 steps "
       "is exceeded by the running average loss over arg2 steps (arg1 >> arg2) by arg3 standard deviations")
-      ->implicit_val("100 10 3.0f");
+      ->implicit_val("1000 10 5.0f");
+  cli.add<std::vector<YAML::Node>>("--custom-fallbacks",
+      "List of custom fallback options after divergence. Each caught divergence exception thrown when --throw-on-divergence conditions are met progresses through another fallback. "
+      "If more exception are caught than fallbacks were specified the process will terminate with an uncaught exception.");
+
   cli.add<bool>("--fp16-fallback-to-fp32",
       "If fp16 training diverges and throws try to continue training with fp32 precision");
+  cli.alias("fp16-fallback-to-fp32", "true", [](YAML::Node& config) {
+    // use default custom-fallbacks to handle DivergenceException for fp16
+    config["custom-fallbacks"] = std::vector<YAML::Node>({ 
+      YAML::Load("{fp16 : false, precision: [float32, float32], cost-scaling: []}")
+     });
+  });
+
+  // @TODO: implement this next:
+  // cli.add<std::string>("--recover-from-fallback-after",
+  //     "Attempt to return to default options once the training has progressed in fallback mode by this many units. "
+  //     "Allowed units are the same as for disp-freq (i.e. (u)pdates, (t)okens, (e)pochs)");
+
   cli.add<size_t>("--gradient-norm-average-window",
       "Window size over which the exponential average of the gradient norm is recorded (for logging and scaling). "
       "After this many updates about 90% of the mass of the exponential average comes from these updates",